Develop

+++ app/Models/Solution.php
@@ -10,6 +10,9 @@ use Illuminate\Support\Facades\DB;
 class Solution extends Model
 {
     use HasFactory;
+    
+    // ブラックリスト
+    protected $guarded = ['id'];
 
     /*
      * 1対多（逆）所属
@@ -182,10 +185,13 @@ class Solution extends Model
 
     /**
      * 解決策削除処理
+     * @param $id
      */
     public function deleteSolution($id)
     {
         $solution = Solution::find($id);
-        $solution->delete();
+        if(Auth::id() === $solution->user_id) {
+            $solution->delete();
+        }
     }
 }
